requirements.txt
argcomplete==1.10.0
argon2-cffi==19.2.0
asgiref==3.2.3
bcrypt==3.1.7
beautifulsoup4==4.8.0
bleach==3.1.0
bootstrap4==0.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Cython==0.29.15
Django==3.0
django-bleach==0.6.1
django-bootstrap4==1.1.1
django-braces==1.13.0
django-clear-cache==0.3
django-debug-toolbar==2.1
django-html-sanitizer==0.1.5
django-parler==2.0
django-six==1.0.4
djangorestframework==3.11.0
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
get==2019.4.13
idna==2.8
image==1.5.28
IMAPClient==2.1.0
lxml==4.5.0
misaka==2.1.1
nltk==3.4.5
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.32
pdfminer.six==20181108
Pillow==6.2.1
post==2019.4.13
promise==2.2.1
public==2019.4.13
pycparser==2.19
pycryptodome==3.9.7
PyPDF2==1.26.0
pystache==0.5.4
pytesseract==0.3.2
python-pptx==0.6.18
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
sqlparse==0.3.0
tesseract==0.1.3
textract==1.6.3
tzlocal==1.5.1
urllib3==1.25.7
webencodings==0.5.1
xlrd==1.2.0
XlsxWriter==1.2.8
